Split component metadata helpers
