+++ components/daikin_ekhhe/daikin_ekhhe.h
@@ -851,92 +851,8 @@ struct SelectFieldSpec {
   uint8_t bit_width;
 };
 
-static const std::map<std::string, NumberFieldSpec> NUMBER_FIELD_SPECS = {
-  {P1_LOW_WAT_PROBE_HYST,   {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_P1_IDX, NumberFieldEncoding::UINT8}},
-  {P2_HEAT_ON_DELAY,        {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_P2_IDX, NumberFieldEncoding::UINT8}},
-  {P3_ANTL_SET_T,           {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_P3_IDX, NumberFieldEncoding::UINT8}},
-  {P4_ANTL_DURATION,        {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_P4_IDX, NumberFieldEncoding::UINT8}},
-  {P7_DEFROST_CYCLE_DELAY,  {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_P7_IDX, NumberFieldEncoding::UINT8}},
-  {P8_DEFR_START_THRES,     {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_P8_IDX, NumberFieldEncoding::INT8}},
-  {P9_DEFR_STOP_THRES,      {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_P9_IDX, NumberFieldEncoding::UINT8}},
-  {P10_DEFR_MAX_DURATION,   {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_P10_IDX, NumberFieldEncoding::UINT8}},
-  {P17_HP_START_DELAY_DIG1, {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_P17_IDX, NumberFieldEncoding::UINT8}},
-  {P18_LOW_WAT_T_DIG1,      {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_P18_IDX, NumberFieldEncoding::UINT8}},
-  {P19_LOW_WAT_T_HYST,      {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_P19_IDX, NumberFieldEncoding::UINT8}},
-  {P20_SOL_DRAIN_THRES,     {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_P20_IDX, NumberFieldEncoding::UINT8}},
-  {P21_LOW_WAT_T_HP_STOP,   {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_P21_IDX, NumberFieldEncoding::UINT8}},
-  {P22_UP_WAT_T_EH_STOP,    {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_P22_IDX, NumberFieldEncoding::UINT8}},
-  {P25_UP_WAT_T_OFFSET,     {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_P25_IDX, NumberFieldEncoding::INT8}},
-  {P26_LOW_WAT_T_OFFSET,    {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_P26_IDX, NumberFieldEncoding::INT8}},
-  {P27_INLET_T_OFFSET,      {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_P27_IDX, NumberFieldEncoding::INT8}},
-  {P28_DEFR_T_OFFSET,       {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_P28_IDX, NumberFieldEncoding::INT8}},
-  {P29_ANTL_START_HR,       {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_P29_IDX, NumberFieldEncoding::UINT8}},
-  {P30_UP_WAT_T_EH_HYST,    {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_P30_IDX, NumberFieldEncoding::UINT8}},
-  {P31_HP_PERIOD_AUTO,      {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_P31_IDX, NumberFieldEncoding::UINT8}},
-  {P32_EH_AUTO_TRES,        {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_P32_IDX, NumberFieldEncoding::UINT8}},
-  {P34_EEV_SH_PERIOD,       {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_P34_IDX, NumberFieldEncoding::UINT8}},
-  {P35_EEV_SH_SETPOINT,     {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_P35_IDX, NumberFieldEncoding::INT8}},
-  {P36_EEV_DSH_SETPOINT,    {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_P36_IDX, NumberFieldEncoding::UINT8}},
-  {P37_EEV_STEP_DEFR,       {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_P37_IDX, NumberFieldEncoding::UINT8}},
-  {P38_EEV_MIN_STEP_AUTO,   {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_P38_IDX, NumberFieldEncoding::UINT8}},
-  {P40_EEV_INIT_STEP,       {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_P40_IDX, NumberFieldEncoding::UINT8}},
-  {P41_AKP1_THRES,          {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_P41_IDX, NumberFieldEncoding::INT8}},
-  {P42_AKP2_THRES,          {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_P42_IDX, NumberFieldEncoding::INT8}},
-  {P43_AKP3_THRES,          {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_P43_IDX, NumberFieldEncoding::INT8}},
-  {P44_EEV_KP1_GAIN,        {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_P44_IDX, NumberFieldEncoding::INT8}},
-  {P45_EEV_KP2_GAIN,        {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_P45_IDX, NumberFieldEncoding::INT8}},
-  {P46_EEV_KP3_GAIN,        {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_P46_IDX, NumberFieldEncoding::INT8}},
-  {P47_MAX_INLET_T_HP,      {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_P47_IDX, NumberFieldEncoding::UINT8}},
-  {P48_MIN_INLET_T_HP,      {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_P48_IDX, NumberFieldEncoding::INT8}},
-  {P49_EVA_INLET_THRES,     {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_P49_IDX, NumberFieldEncoding::UINT8}},
-  {P50_ANTIFREEZE_SET,      {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_P50_IDX, NumberFieldEncoding::UINT8}},
-  {P51_EVA_HIGH_SET,        {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_P51_IDX, NumberFieldEncoding::UINT8}},
-  {P52_EVA_LOW_SET,         {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_P52_IDX, NumberFieldEncoding::UINT8}},
-  {P54_LOW_PRESS_BYPASS,    {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_P54_IDX, NumberFieldEncoding::UINT8}},
-  {ECO_T_TEMPERATURE,       {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_ECO_TTARGET_IDX, NumberFieldEncoding::UINT8}},
-  {AUTO_T_TEMPERATURE,      {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_AUTO_TTARGET_IDX, NumberFieldEncoding::UINT8}},
-  {BOOST_T_TEMPERATURE,     {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_BOOST_TTGARGET_IDX, NumberFieldEncoding::UINT8}},
-  {ELECTRIC_T_TEMPERATURE,  {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_ELECTRIC_TTARGET_IDX, NumberFieldEncoding::UINT8}},
-  {VAC_DAYS,                {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_VAC_DAYS, NumberFieldEncoding::UINT8}},
-
-  {P53_EVA_FAN_DEFR_SPEED,  {WritablePacketFamily::EXTENDED, DaikinEkhheComponent::EXT_PACKET_P53_IDX, NumberFieldEncoding::UINT8}},
-  {P55_EVA_BAND1_PROP,      {WritablePacketFamily::EXTENDED, DaikinEkhheComponent::EXT_PACKET_P55_IDX, NumberFieldEncoding::UINT8}},
-  {P56_EVA_MAX_ACT_DELTA,   {WritablePacketFamily::EXTENDED, DaikinEkhheComponent::EXT_PACKET_P56_IDX, NumberFieldEncoding::UINT8}},
-  {P57_EVA_MAX_DEACT_DELTA, {WritablePacketFamily::EXTENDED, DaikinEkhheComponent::EXT_PACKET_P57_IDX, NumberFieldEncoding::UINT8}},
-  {P59_EVA_FAN_OFF_SPEED,   {WritablePacketFamily::EXTENDED, DaikinEkhheComponent::EXT_PACKET_P59_IDX, NumberFieldEncoding::UINT8}},
-  {P60_EVA_AIR_DELTA1,      {WritablePacketFamily::EXTENDED, DaikinEkhheComponent::EXT_PACKET_P60_IDX, NumberFieldEncoding::UINT8}},
-  {P61_EVA_AIR_DELTA2,      {WritablePacketFamily::EXTENDED, DaikinEkhheComponent::EXT_PACKET_P61_IDX, NumberFieldEncoding::UINT8}},
-  {P62_EVA_AIR_DELTA3,      {WritablePacketFamily::EXTENDED, DaikinEkhheComponent::EXT_PACKET_P62_IDX, NumberFieldEncoding::UINT8}},
-  {P63_EVA_AIR_DELTA4,      {WritablePacketFamily::EXTENDED, DaikinEkhheComponent::EXT_PACKET_P63_IDX, NumberFieldEncoding::UINT8}},
-  {P64_EVA_AIR_DELTA5,      {WritablePacketFamily::EXTENDED, DaikinEkhheComponent::EXT_PACKET_P64_IDX, NumberFieldEncoding::UINT8}},
-  {P65_EVA_AIR_DELTA6,      {WritablePacketFamily::EXTENDED, DaikinEkhheComponent::EXT_PACKET_P65_IDX, NumberFieldEncoding::UINT8}},
-  {P66_EVA_BAND2_PROP,      {WritablePacketFamily::EXTENDED, DaikinEkhheComponent::EXT_PACKET_P66_IDX, NumberFieldEncoding::UINT8}},
-  {P67_EVA_BAND3_PROP,      {WritablePacketFamily::EXTENDED, DaikinEkhheComponent::EXT_PACKET_P67_IDX, NumberFieldEncoding::UINT8}},
-  {P68_EVA_BAND4_PROP,      {WritablePacketFamily::EXTENDED, DaikinEkhheComponent::EXT_PACKET_P68_IDX, NumberFieldEncoding::UINT8}},
-  {P69_EVA_BAND5_PROP,      {WritablePacketFamily::EXTENDED, DaikinEkhheComponent::EXT_PACKET_P69_IDX, NumberFieldEncoding::UINT8}},
-  {P70_EVA_BAND6_PROP,      {WritablePacketFamily::EXTENDED, DaikinEkhheComponent::EXT_PACKET_P70_IDX, NumberFieldEncoding::UINT8}},
-  {P71_EC_FAN_SILENT_RED,   {WritablePacketFamily::EXTENDED, DaikinEkhheComponent::EXT_PACKET_P71_IDX, NumberFieldEncoding::UINT8}},
-  {P72_EC_FAN_REG_GAIN,     {WritablePacketFamily::EXTENDED, DaikinEkhheComponent::EXT_PACKET_P72_IDX, NumberFieldEncoding::UINT8}},
-};
-
-static const std::map<std::string, SelectFieldSpec> SELECT_FIELD_SPECS = {
-  {OPERATIONAL_MODE,     {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_MODE_IDX, BIT_POSITION_NO_BITMASK, 1}},
-  {P12_EXT_PUMP_MODE,    {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_P12_IDX, BIT_POSITION_NO_BITMASK, 1}},
-  {P14_EVA_BLOWER_TYPE,  {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_P14_IDX, BIT_POSITION_NO_BITMASK, 1}},
-  {P16_SOLAR_MODE_INT,   {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_P16_IDX, BIT_POSITION_NO_BITMASK, 1}},
-  {P23_PV_MODE_INT,      {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_P23_IDX, BIT_POSITION_NO_BITMASK, 1}},
-  {P24_OFF_PEAK_MODE,    {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_P24_IDX, BIT_POSITION_NO_BITMASK, 1}},
-  {P58_EVA_FAN_COMP_OFF, {WritablePacketFamily::EXTENDED, DaikinEkhheComponent::EXT_PACKET_P58_IDX, BIT_POSITION_NO_BITMASK, 1}},
-
-  {POWER_STATUS,          {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_MASK1_IDX, 0, 1}},
-  {P39_EEV_MODE,          {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_MASK1_IDX, 2, 1}},
-  {P13_HW_CIRC_PUMP_MODE, {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_MASK1_IDX, 4, 1}},
-  {P11_DISP_WAT_T_PROBE,  {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_MASK2_IDX, 0, 1}},
-  {P15_SAFETY_SW_TYPE,    {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_MASK2_IDX, 1, BIT_WIDTH_P15_STEPPED}},
-  {P5_DEFROST_MODE,       {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_MASK2_IDX, 2, 1}},
-  {P6_EHEATER_DEFROSTING, {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_MASK2_IDX, 3, 1}},
-  {P33_EEV_CONTROL,       {WritablePacketFamily::MAIN, DaikinEkhheComponent::CC_PACKET_MASK2_IDX, 4, 1}},
-};
+extern const std::map<std::string, NumberFieldSpec> NUMBER_FIELD_SPECS;
+extern const std::map<std::string, SelectFieldSpec> SELECT_FIELD_SPECS;
 
 
 }  // namespace daikin_ekkhe
